+++ generator/test/error_test_data/double_primary_key.dart
@@ -1,0 +1,17 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'double_primary_key.g.dart';
+
+@RealmModel()
+class _Bad {
+  @PrimaryKey()
+  late int first;
+
+  @PrimaryKey()
+  late String second;
+
+  late String another;
+
+  @PrimaryKey()
+  late String third;
+}