+++ generator/test/error_test_data/not_an_indexable_type.dart
@@ -1,0 +1,10 @@
+import 'dart:ffi';
+import 'package:realm_common/realm_common.dart';
+
+//part 'not_an_indexable_type.g.dart';
+
+@RealmModel()
+class _Bad {
+  @Indexed()
+  late Double notAnIndexableType;
+}