+++ generator/test/good_test_data/pinhole.dart
@@ -1,0 +1,8 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'pinhole.g.dart';
+
+@RealmModel()
+class _Foo {
+  int x = 0;
+}