+++ CHANGELOG.md
@@ -4,10 +4,10 @@ x.x.x Release notes (yyyy-MM-dd)
 **This project is in the Alpha stage. All API's might change without warning and no guarantees are given about stability. Do not use it in production.**
 
 ### Enhancements
-* None
+* Support result value from write transaction callbacks ([#294](https://github.com/realm/realm-dart/pull/294/))
 
 ### Fixed
-* Fix running package commands.
+* None
 
 ### Compatibility
 * Dart ^2.15 on Windows, MacOS and Linux